requirements.txt
requests==2.24.0
Flask==1.1.2
PyExecJS==1.5.1
pandas==1.1.3
pycryptodome==3.12.0
